fixed bugs

+++ app/src/main/java/com/example/smartglove/Flex.java
@@ -1,9 +1,9 @@
 package com.example.smartglove;
 
 public class Flex {
-    public double thumb, pointer, middle, ring, pinkie;
+    public int thumb, pointer, middle, ring, pinkie;
 
-    public Flex(double thumb, double pointer, double middle, double ring, double pinkie) {
+    public Flex(int thumb, int pointer, int middle, int ring, int pinkie) {
         this.thumb = thumb;
         this.pointer = pointer;
         this.middle = middle;
